+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `whisk` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://github.com/AldaronLau/semver).
 
-## [0.8.0] - Unreleased
+## [0.8.0] - 2022-09-28
 ### Added
  - `Chan`, `Stream`, `WeakChan`, `WeakStream` type aliases